Update to version 0.0.2
fix: Update DataInputAssociation.cs

In the DataInputAssociation class of BpmnElements namespace, update the way SourceRef and TargetRef attributes are retrieved.

fix: Update BpmnProcessReader.cs

Instead of using TakeWhile to limit the selection based on the presence of the "id" attribute, use Where to filter elements that have the "id" attribute.